Refactor search service, location search
