+++ src/app/services/services.module.ts
@@ -9,6 +9,7 @@ import { ScrollService } from './scroll.service';
 import { AnalyticsService } from './analytics.service';
 import { RoutingService } from './routing.service';
 import { Ng2PageScrollModule } from 'ng2-page-scroll';
+import { SearchService } from './search.service';
 
 @NgModule({
   imports: [
@@ -24,7 +25,8 @@ export class ServicesModule {
         PlatformService,
         LoadingService,
         AnalyticsService,
-        RoutingService
+        RoutingService,
+        SearchService
       ]
     };
   }
